+++ Assets/Scripts/HitPoints/IDamageReceiver.cs
@@ -2,6 +2,6 @@
 
     public interface IDamageReceiver
     {
-        public UnityAction<float> OnApplyDamage { get; set; }
-        public void ApplyDamage(float damageValue);
+        public UnityAction<int> OnApplyDamage { get; set; }
+        public void ApplyDamage(int damageValue);
     }
